+++ src/components/__tests__/CustomListEditor-test.tsx
@@ -246,7 +246,7 @@ describe("CustomListEditor", () => {
     getEntriesStub.restore();
   });
 
-  it("shouldn't allow you to save unless the list has a title", () => {
+  it("shouldn't allow you to save unless the list has a title and entries", () => {
     wrapper = mount(
       <CustomListEditor
         library={library}
@@ -267,6 +267,14 @@ describe("CustomListEditor", () => {
     expect(saveButton.props().disabled).to.equal(true);
 
     wrapper.setState({ title: "new list title" });
+    saveButton = wrapper.find(".save-or-cancel-list").find(Button).at(0);
+
+    expect(saveButton.props().disabled).to.equal(true);
+
+    wrapper.setState({
+      title: "new list title",
+      entries: [{ id: "1234", title: "a", authors: [] }],
+    });
     saveButton = wrapper.find(".save-or-cancel-list").find(Button).at(0);
 
     expect(saveButton.props().disabled).to.equal(false);
@@ -304,7 +312,7 @@ describe("CustomListEditor", () => {
       "getEntries"
     ).returns(newEntries);
     const saveButton = wrapper.find(".save-or-cancel-list").find(Button).at(0);
-    wrapper.setState({ title: "new list title" });
+    wrapper.setState({ title: "new list title", entries: newEntries });
     saveButton.simulate("click");
 
     expect(editCustomList.callCount).to.equal(1);
@@ -718,21 +726,25 @@ describe("CustomListEditor", () => {
       expect(hasChanges).to.equal(true);
     });
   });
-  it("should know whether the list title is blank", () => {
-    // There is a list property with a title
-    expect(wrapper.instance().isTitleEmpty()).to.be.false;
+  it("should know whether the list title is blank, or there are no entries", () => {
+    // There is a list property with a title and entries
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.false;
     wrapper.setProps({ list: null });
+    wrapper.setState({ entries: [] });
     wrapper.setState({ title: null });
     // New list, no title
-    expect(wrapper.instance().isTitleEmpty()).to.be.true;
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.true;
     // New list, title is still just the placeholder
     wrapper.setState({ title: "list title" });
-    expect(wrapper.instance().isTitleEmpty()).to.be.true;
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.true;
     // New list, placeholder has been deleted.
     wrapper.setState({ title: "" });
-    expect(wrapper.instance().isTitleEmpty()).to.be.true;
-    // Adding a title...
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.true;
+    // Adding a title, but no entries...
     wrapper.setState({ title: "testing..." });
-    expect(wrapper.instance().isTitleEmpty()).to.be.false;
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.true;
+    // Adding entries...
+    wrapper.setState({ entries: [{ id: "1234", title: "a", authors: [] }] });
+    expect(wrapper.instance().isTitleOrEntriesEmpty()).to.be.false;
   });
 });