+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## Changelog
 
-### 08/30/21 (to be included in future release)
+### v0.5.5
 
 #### Updated
 
+- Updated `CustomListEditor` so that the "Save this list" button is disabled if either title or entries are absent, and enabled if both are present.
 - Updated the book cover editor so that the current cover URL pulls through not as a value in the input (as described in v0.5.3), but as a description beneath the input.
 
 ### v0.5.4